SSH-related tweaks
Looks like MOTD runs as root on a regular Ubuntu system (on WSL2 it runs
as the user defined in /etc/wsl.conf), so not really useful to output
the "host" element (i.e. "root@host") in screenfetch. Also, perhaps not
a good idea to symlink to file from the Git-repository (as it runs as
root upon login)...

Furthermore, show the hostname in Powerline-Go for SSH sessions only.