Include WSL-init in .bashrc.d
Used to be run via Windows Task Scheduler at system boot, but every now
and then this crashed WSL (either not doing the init or, even worse,
requiring a manual "wsl --shutdown" to revive WSL again).

Since this is just convenience stuff (and not starting any critical
services) it's just as good to start it the first time I open my shell.