Replace --exit-with-session with --exit-with-x11
On closer inspection of the manual, it appears one should not use
--exit-with-session, but instead should use --exit-with-x11 to achieve
the desired effect.
Found this out the hard way when trying to type something on the
terminal (which didn't work that well) after pulling up a D-Bus session
towards a broken X server

Also, hide the error that occurs when the X server (i.e. my laptop)
isn't available.