Tweaks for Windows 11 (21H2 – build 22000)
With the introduction of WSLg, there's no need to run a custom X11
server (VcXsrv) anymore. As such the WSL-specific X11 initialisation
isn't required anymore either. A single script the runs both for local
and remote sessions now suffices.

Additionally, the latest version of WSL2 did introduce a new issue/bug
with regards to a custom /etc/resolv.conf; there's a comment in the file
indication how to work around the issue.