Use wsl.conf to start services
Instead of starting them as part of .bashrc, use the "[boot] command"
option in wsl.conf to bring them up whenever the WSL2 VM is started.
The services started are mainly used to support some of my use cases in
Bash, so the impact of this change is minor — nonetheless this seems
a more proper way of doing things.

When the systemd support in WSL2 matures this will be unnecessary and
the same effect can simply be achieved through the (default) systemd
configuration...