Tweak the pyenv update-check
Instead of having the scheduled task delete the stale update
notification file, have the bashrc-script do it. This way the (heavy)
update-check can be run infrequently (ie, once per week) and it doesn't
need all kinds of logic to "sometimes run the update-check, but always
remove the stale notification file"... Much better to have the bashrc-
script (which always needs to run anyway) do it.