Add usbip/Yubikey fallback option
Toying around with WSL2's new mirrored networking-mode surfaced some
issues with usbipd-win. Hopefully these'll be addressed soon. For now a
short-term workaround is required.

Instead of reinstating the previous solution, I've added fallback logic
that takes the host's IP address in the Hyper-V virtual network in case
no $USBIP_HOST is specified.

This is a middle ground between the previous solution (with hard-coded
local IP addresses and as strict as possible firewall setup) and the
future solution (through the loopback interface). This one exposes
usbipd-win on the entire 172.16.0.0/12-range (which Hyper-V seems to use
for its virtual network), not optimal, but better than exposing it on
whatever IP address (and thus local network) my laptop happens to be on
when using it...