Small improvements to the install-procedure
It's now possible to run a single install-part using install.sh (instead
of calling it directly, which only works for a small number of the parts
due to them relying on variables set by install.sh).

On Ubuntu 22.04, "50-landscape-sysinfo" went from an actual file to a
symlink, so "chmod -x" wasn't effective anymore.