Use native go(lang) instead of prebuilt bins
Turns out to be easier and more flexible; especially now that Go itself
can be installed as a snap on WSL2 (on Ubuntu with systemd enabled).

Required the powerline-go initialisation to be moved all the way to the
end of the chain (more sensible location to begin with actually).

This in turn necessitated another change: The "_powerline_ps1"
PROMPT_COMMAND entry should be _appended_ to prevent issues with
direnv's entry. direnv Would now get executed _after_ powerline_go
which lead to the powerline not immediately picking up the changes made
by direnv...