Expand (automatic) installations of packages
Instead of listing a handful of commands in install/README.md, put them
in install scripts. The commands can still be executed manually if
you're so inclined...

Move from using the apt version of gcloud CLI to using its snap. The
apt-package is a monstrosity (takes five-minutes to compile/install),
snap to the rescue!

Update install.sh to always execute the single script passed in as its
argument (used to only do that if the referenced script was executable).

Rename "wget.d" to a more generic "bins.d" (still not optimally happy
with the name, but alas) and include a script to install "TODO get", a
tool to find "TODO" comments in code.
As I am in the habit of using "XXX" to indicate stuff that still needs
to be fixed, there's a small (inline) patch in the script to update the
Go code to include that marker prior to compilation (the is also the
primary motivator for switching from precompiled Go-binaries to
installing Go itself 😇).

The final install-script ("93-install") automatically sources all
(executable) files from the "bins.d" folder, tying everything together.