Fix (and improve upon 🚀) the linter
It's tied a little bit too closely into the repository itself... 😇
Never mind about that, but did make some other improvements: It now
explicitly specifies which versions of Node, Python, and (newly added)
Go to use.

It further more explicitly adds "~/.local/bin" to the PATH. Was already
enforced (ie, linter would fail if this was not the case), but not
automatically corrected. This caused local testing (with "gh act") to
fail, so we now always add it to the PATH regardless... Probably means
it's added twice in GitHub runs, but that's a problem for another day.