Switch from neofetch to fastfetch
Faster, better support for WSL2 and under active development. Only
drawback is the somewhat more complicated installation procedure.

Also, note the update to Prettier's configuration to _disallow_ trailing
commas in JSONC files (which is Microsoft's recommendation). There's an
ongoing discussion on Prettier's GitHub on how to handle this properly
long-term...