Clean up ".config" somewhat
Instead of having a duplicate powerline-go configuration with an
absolute minimal difference (two paths that only exist on WSL, but don't
do any harm on the RPis), use a single one everywhere.

Furthermore, don't keep the htop configuration in the repo (it actually
differs quite a bit between various distros/version, and depending on
the amount of cores in a system). Instead, keep a backup copy of WSL's
version in OneDrive (and include btop's too).