Symlink SSH configuration from OneDrive
Now that my D-drive is mounted with a proper fmask and metadata, it's
possible to symlink the SSH configuration from OneDrive (instead of
just backing it up there via rclone).

This is more convenient (as from time to time it's easy to modify it
from Windows) and more consistent (as Windows' SSH configuration is
also symlinked from OneDrive).