Proper split between .bashrc and .profile
Instead of dumping everything in .bashrc.d, move all scripts that merely
set environment variables to .profile.d and source them from profile
instead.

This way, a non-interactive (and non-Bash) shell is still tooled up with
a proper environment. As a consequence, all scripts in .profile.d are
now sh (ie, POSIX-compliant) instead of Bash, as .profile can be sourced
from non-Bash shells too.

On a day-to-day basis in WSL2 this doesn't do a whole lot, but the
impact is visible in VS Code which sources .profile once for the entire
session and subsequently only sources .bashrc when a terminal instance
is opened. Some "convincing" was required in .profile to ensure things
are loaded in the proper order (and not loaded twice).