Fix (and improve) install.sh
The parts-directory is now actually optional. Temporary directories
created for "bins.d"-parts are now also removed when the installation
exits with an error.

In case of error, part-output is now properly displayed (and the exit
code set to 1); the remaining parts are still executed (time will tell
if that's actually helpful or not).